Fix issue of stuck on waiting for bind adapter when add group config
The nng_dial function sometimes returned after call the callback
function manager_bind_adapter by nng_pipe_notify. This causes the
manager_bind_adapter function to be called with another missing adapter,
the bind_count of current bound adapter should not be changed, so it
will be stuck on waiting for the bind adapter.